6309152 gcc and mdesc don't get along 6309756 gcc and niagara perfctr don't get along 6309977 gcc and glvc don't get along 6310446 gcc and gcn don't get along 6310609 gcc and sun4v kernel support don't get along

+++ usr/src/common/mdesc/mdesc_init_intern.c
@@ -45,7 +45,7 @@ md_init_intern(uint64_t *ptr, void *(*allocp)(size_t),
 	 * Very basic checkup for alignment to avoid
 	 * bus error issues.
 	 */
-	if ((((uint64_t)ptr)&7) != 0)
+	if ((((uintptr_t)ptr)&7) != 0)
 		return (NULL);
 
 	mdp = (md_impl_t *)allocp(sizeof (md_impl_t));
